third_party/blink/web_tests/http/tests/eventsource/eventsource-content-type-charset-expected.txt

CONSOLE ERROR: EventSource's response has a charset ("windows-1251") that is not UTF-8. Aborting the connection.
CONSOLE ERROR: EventSource's response has a MIME type ("text/event-stream-foobar") that is not "text/event-stream". Aborting the connection.
Test EventSource with an event-stream with a Content-Type with a charset is still recognized.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3DUTF-8") did not throw exception.
Content type: text/event-stream; charset=UTF-8
PASS !!es.sawError is false
PASS es.sawMessage is true
PASS es.sawOpen is true
PASS es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3Dwindows-1251") did not throw exception.
Content type: text/event-stream; charset=windows-1251
PASS es.sawError is true
PASS !!es.sawMessage is false
PASS !!es.sawOpen is false
PASS es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3Dutf-8") did not throw exception.
Content type: text/event-stream; charset=utf-8
PASS !!es.sawError is false
PASS es.sawMessage is true
PASS es.sawOpen is true
PASS es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream%3B%20charset%3D%22UTF-8%22") did not throw exception.
Content type: text/event-stream; charset="UTF-8"
PASS !!es.sawError is false
PASS es.sawMessage is true
PASS es.sawOpen is true
PASS es = new EventSource("/eventsource/resources/response-content-type-charset.php?contentType=text/event-stream-foobar%3B") did not throw exception.
Content type: text/event-stream-foobar;
PASS es.sawError is true
PASS !!es.sawMessage is false
PASS !!es.sawOpen is false
PASS successfullyParsed is true

TEST COMPLETE
